update for duckdb tpcc 4

+++ duckdb_lab/types.h
@@ -2,5 +2,5 @@
 #include <stdint.h>
 #include <vector>
 
-using ID_TYPE = int64_t;
+using ID_TYPE = int;
 using IDS_TYPE = std::vector<ID_TYPE>;
